Project Paper
The medical sector is one of the largest sectors in Bangladesh. In Bangladesh, 70 percent
of people live in rural areas. They don’t know the hospital cost and test system.
Whenever they come to Dhaka for hospital service, they face a lot of trouble. They don’t
even know which hospital is best for them according to their ability. That’s why we
develop an application in which the format is .exe. It can be run on any windows-based
computer. Now almost all have a personal computer. It is one of the daily useable things
in our life. In any computer, anyone can run this on his/her computer. With this app, one
can easily know all the information that he/she needs. Like he can know the test rate, bed
rate, ICU rate, doctor’s information, and their visitor hour of most of the hospitals in
Dhaka city. All the information in just a click away. It reduces hassle. The
emergency patient can be admitted within a short time. Our main target is to achieve the
best output from hospital service. Our app will give the basic knowledge of hospital costs
and other hospital-related information.